Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    float f;
-    scanf("%f",&f);
-    printf("You entered: %.4f",f);
+    double f;
+    scanf("%lf",&f);
+    printf("You entered: %.4lf",f);
     return 0;
 }
